Fix #115930: custom colored bones are invisible when drawn in stick mode
The fix here is simple: we ignore the alpha channel in the bone theme
colors, setting it unconditionally to fully opaque.  This is correct
because conceptually the bone theme colors are just RGB, not RGBA,
and the alpha channel doesn't always get set consistently.

Pull Request: https://projects.blender.org/blender/blender/pulls/121256